+++ NeuralNetwork/LossFunctions.h
@@ -14,6 +14,6 @@ struct LossFunctions
 {
 	static Matrix<double> squared_error(const Matrix<double>& y, const Matrix<double>& y_hat);
 	static Matrix<double> squared_error_derived(const Matrix<double>& y, const Matrix<double>& y_hat);
-	static double mean_squared_error(const Matrix<double>& errors);
+	static long double mean_squared_error(const Matrix<double>& errors);
 };
 #endif
